+++ src/entities/HomotopyDensity.jl
@@ -20,24 +20,22 @@ Minor eigenvectors are sometimes called "trailing eigenvectors," or "residual mo
 @kwdef struct HomotopyDensity{
   H <: HomotopyRepresentation, # serde friendly when using DFG.statekind representation, but also supports Manifolds.jl direclty
   P <: AbstractArray, # serde relies on DFG statekind mechanism, does not guarantee serde when directly using Manifolds wo DFG.statekind
-  # HL,  # FIXME, deprecating
   HT,  # FIXME, deprecating
   MD # Use only easy to JSON.jl lift lower serde -- e.g. Dict{Int, Vector{Float64}} when storing just diagonal covariances for leaves of tree, or similar
 }
     representationkind::H
     observability::Vector{Float64} = zeros(manifold_dimension(getManifold(representationkind)))
-    elements::Vector{P}
-    weights::Vector{Float64} = Vector{Float64}(ones(length(elements))) ./ length(elements)  # TODO rename to mixture_weights
+    points::Vector{P}
+    weights::Vector{Float64} = Vector{Float64}(ones(length(points))) ./ length(points)  # TODO rename to mixture_weights
     """ 
-    Geometric elements permute field, allows fast binary tree operations and geometric elements splits for manellic (ball) trees. 
-    - Geometric split reqs at least 2*(N+1)-1 elements -- e.g. when nodes have only right children, elements=[1,2,-3].
+    Geometric points permute field, allows fast binary tree operations and geometric points splits for manellic (ball) trees. 
+    - Geometric split reqs at least 2*(N+1)-1 points -- e.g. when nodes have only right children, points=[1,2,-3].
     """
     structure::SparseArrays.SparseVector{Vector{Int}, Int} = SparseArrays.sparsevec(
-      Dict(1 => collect(1:length(elements))), 
-      5*(length(elements)) # large buffer space where impact on resources mitigated via sparsevec
+      Dict(1 => collect(1:length(points))), 
+      5*(length(points)) # large buffer space where impact on resources mitigated via sparsevec
     )
     tree_kernels::Vector{HT}        # TODO rename to leading
-    # leaf_kernels::Vector{HL}        # FIXME, remove
     """ 
     Store minor details such as leaf bandwidth or eigenvectors associated with minor eigenvalues.
     - When lifted for compute efficiency, this field is likely to hold something like PDMats.