+++ test/manellic/testManifoldTreeConstr.jl
@@ -551,7 +551,7 @@ end
 
     mtree = HomotopyDensity_legacy(;
         manifold = M,
-        elements = pts,
+        points = pts,
         # leaf_kernels = lkern,                           # leaf_kernels
         tree_kernels = Vector{KT}(undef, N),       # tree_kernels
     );